Refactor of caf-api for 9.0
The goals of this change are to improve clarity of the interfaces, simplify the api package so it more strictly
adheres to being an API alone (with little if any logic), and standardise on the use of interfaces and base
classes.

All entities in caf-api now have an interface which can be implemented directly without any overriding logic.
Logic for components like ConfigurationSource has moved to CafConfigurationSource in the config-caf package.
Logic for Worker and DefaultWorkerFactory has moved to the worker-caf package, and the classes have become
AbstractWorker and AbstractWorkerFactory respectively.

The classes related to service naming (Name and ServicePath) have moved to util-naming.

The goal is to use default methods in the interfaces going forward to avoid breaking changes where possible.